+++ docs/changelog.md
@@ -17,10 +17,11 @@ to [Anselm Hahn](https://github.com/Anselmoo) for these two contributions!
 - The examples in the documentation are now executed as part of the test suite to increase the coverage.
 
 **Changed**
-- We have changed the default table to `style = "table-layout:auto"` to fix an issue on the width of index columns (default `style` was `width:auto` previously) ([#130](https://github.com/mwouts/itables/issues/130))
+- We have changed the default table to `style = "table-layout:auto;width:auto;margin:auto"` to fix an issue on the width of index columns (default `style` was `width:auto` previously) ([#130](https://github.com/mwouts/itables/issues/130))
 - The default classes applied to datatables are now `["display", "nowrap"]`
 - We have changed the default order to `order = []` i.e. we don't sort anymore the table, even when the index is monotonic, to fix an issue in the order of categories ([#135](https://github.com/mwouts/itables/issues/135))
 - We have set an explicit `maxRows = 0` and also increased `maxColumns` to `200` (instead of Pandas' default at 20).
+
 
 1.3.5 (2022-11-12)
 ------------------